Rework whole muc edit ui to be more modern
