+++ Monal/Classes/ContactDetails.swift
@@ -6,14 +6,11 @@
 //  Copyright © 2021 Monal.im. All rights reserved.
 //
 
-import UIKit
-import SwiftUI
-import monalxmpp
-
 struct ContactDetails: View {
     var delegate: SheetDismisserProtocol
     private var account: xmpp
-    private var isGroupModerator = false
+    private var ownRole: String
+    private var ownAffiliation: String
     @StateObject var contact: ObservableKVOWrapper<MLContact>
     @State private var showingBlockContactConfirmation = false
     @State private var showingCannotBlockAlert = false
@@ -24,6 +21,15 @@ struct ContactDetails: View {
     @State private var showingCannotEncryptAlert = false
     @State private var showingShouldDisableEncryptionAlert = false
     @State private var isEditingNickname = false
+    @State private var inputImage: UIImage?
+    @State private var showingImagePicker = false
+    @State private var showingSheetEditSubject = false
+    @State private var showingDestroyConfirmation = false
+    @State private var alertPrompt = AlertPrompt(dismissLabel: Text("Close"))
+    @State private var showAlert = false
+    @State private var success = false
+    @State private var successCallback: monal_void_block_t?
+    @StateObject private var overlay = LoadingOverlayState()
 
     init(delegate: SheetDismisserProtocol, contact: ObservableKVOWrapper<MLContact>) {
         self.delegate = delegate
@@ -31,24 +37,148 @@ struct ContactDetails: View {
         self.account = MLXMPPManager.sharedInstance().getConnectedAccount(forID: contact.accountId)!
 
         if contact.isGroup {
-            let ownRole = DataLayer.sharedInstance().getOwnRole(inGroupOrChannel: contact.obj) ?? "none"
-            self.isGroupModerator = (ownRole == "moderator")
+            self.ownRole = DataLayer.sharedInstance().getOwnRole(inGroupOrChannel: contact.obj) ?? "none"
+            self.ownAffiliation = DataLayer.sharedInstance().getOwnAffiliation(inGroupOrChannel:contact.obj) ?? "none"
+        } else {
+            self.ownRole = "none"
+            self.ownAffiliation = "none"
         }
     }
 
+    private func errorAlert(title: Text, message: Text = Text("")) {
+        alertPrompt.title = title
+        alertPrompt.message = message
+        showAlert = true
+    }
+    
+    private func successAlert(title: Text, message: Text = Text("")) {
+        alertPrompt.title = title
+        alertPrompt.message = message
+        showAlert = true
+        self.success = true // < dismiss entire view on close
+    }
+    
     var body: some View {
         Form {
             Section {
-                ContactDetailsHeader(delegate:delegate, contact:contact)
+                VStack(spacing: 20) {
+                    Image(uiImage: contact.avatar)
+                        .resizable()
+                        .scaledToFit()
+                        .applyClosure {view in
+                            if contact.isGroup {
+                                if ownAffiliation == "owner" {
+                                    view.accessibilityLabel((contact.mucType == "group") ? "Change Group Avatar" : "Change Channel Avatar")
+                                        .onTapGesture {
+                                            showingImagePicker = true
+                                        }
+                                } else {
+                                    view.accessibilityLabel((contact.mucType == "group") ? "Group Avatar" : "Channel Avatar")
+                                }
+                            } else {
+                                view.accessibilityLabel("Avatar")
+                            }
+                        }
+                        .frame(width: 150, height: 150, alignment: .center)
+                        .shadow(radius: 7)
+                        .sheet(isPresented:$showingImagePicker) {
+                            ImagePicker(image:$inputImage)
+                        }
+                    
+                    
+                    Button {
+                        UIPasteboard.general.setValue(contact.contactJid as String, forPasteboardType:UTType.utf8PlainText.identifier as String)
+                        UIAccessibility.post(notification: .announcement, argument: "JID Copied")
+                    } label: {
+                        HStack {
+                            Text(contact.contactJid as String)
+                            
+                            Image(systemName: "doc.on.doc")
+                                .foregroundColor(.primary)
+                                .accessibilityHidden(true)
+                        }
+                        .accessibilityHint("Copies JID")
+                    }
+                    .buttonStyle(.borderless)
+                        
+                    
+                    //only show account jid if more than one is configured
+                    if MLXMPPManager.sharedInstance().connectedXMPP.count > 1 && !contact.isSelfChat {
+                        Text("Account: \(account.connectionProperties.identity.jid)")
+                    }
+                    
+                    if !contact.isSelfChat && !contact.isGroup {
+                        if let lastInteractionTime = contact.lastInteractionTime as Date? {
+                            if lastInteractionTime.timeIntervalSince1970 > 0 {
+                                Text(String(format: NSLocalizedString("Last seen: %@", comment: ""),
+                                    DateFormatter.localizedString(from: lastInteractionTime, dateStyle: DateFormatter.Style.short, timeStyle: DateFormatter.Style.short)))
+                            } else {
+                                Text(String(format: NSLocalizedString("Last seen: %@", comment: ""), NSLocalizedString("now", comment: "")))
+                            }
+                        } else {
+                            Text(String(format: NSLocalizedString("Last seen: %@", comment: ""), NSLocalizedString("unknown", comment: "")))
+                        }
+                    }
+                    
+                    if !contact.isGroup && (contact.statusMessage as String).count > 0 {
+                        VStack {
+                            Text("Status message:")
+                            Text(contact.statusMessage as String)
+                                .fixedSize(horizontal: false, vertical: true)
+                        }
+                    }
+                    
+                    if contact.isGroup && ((contact.groupSubject as String).count > 0 || ownRole == "moderator") {
+                        VStack {
+                            if ownRole == "moderator" {
+                                Button {
+                                    showingSheetEditSubject.toggle()
+                                } label: {
+                                    if contact.obj.mucType == "group" {
+                                        HStack {
+                                            Text("Group subject:")
+                                            Spacer().frame(width:8)
+                                            Image(systemName: "pencil")
+                                                .foregroundColor(.primary)
+                                                .accessibilityHidden(true)
+                                        }
+                                        .accessibilityHint("Edit Group Subject")
+                                    } else {
+                                        HStack {
+                                            Text("Channel subject:")
+                                            Spacer().frame(width:8)
+                                            Image(systemName: "pencil")
+                                                .foregroundColor(.primary)
+                                                .accessibilityHidden(true)
+                                        }
+                                        .accessibilityHint("Edit Channel Subject")
+                                    }
+                                }
+                                .buttonStyle(.borderless)
+                                .sheet(isPresented: $showingSheetEditSubject) {
+                                    LazyClosureView(EditGroupSubject(contact: contact))
+                                }
+                            } else {
+                                Text("Group subject:")
+                            }
+                            
+                            Text(contact.groupSubject as String)
+                                .fixedSize(horizontal: false, vertical: true)
+                        }
+                    }
+                }
+                .foregroundColor(.primary)
+                .padding([.top, .bottom])
+                .frame(maxWidth: .infinity)
             }
                 
             // info/nondestructive buttons
             Section {
                 Button {
-                    if(contact.isGroup) {
-                        if(!contact.isMuted && !contact.isMentionOnly) {
+                    if contact.isGroup {
+                        if !contact.isMuted && !contact.isMentionOnly {
                             contact.obj.toggleMentionOnly(true)
-                        } else if(!contact.isMuted && contact.isMentionOnly) {
+                        } else if !contact.isMuted && contact.isMentionOnly {
                             contact.obj.toggleMentionOnly(false)
                             contact.obj.toggleMute(true)
                         } else {
@@ -59,14 +189,14 @@ struct ContactDetails: View {
                         contact.obj.toggleMute(!contact.isMuted)
                     }
                 } label: {
-                    if(contact.isMuted) {
+                    if contact.isMuted  {
                         Label {
                             contact.isGroup ? Text("Notifications disabled") : Text("Contact is muted")
                         } icon: {
                             Image(systemName: "bell.slash.fill")
                                 .foregroundColor(.red)
                         }
-                    } else if(contact.isGroup && contact.isMentionOnly) {
+                    } else if contact.isGroup && contact.isMentionOnly {
                         Label {
                             Text("Notify only when mentioned")
                         } icon: {
@@ -83,9 +213,9 @@ struct ContactDetails: View {
                 }
                 
 #if !DISABLE_OMEMO
-                if((!contact.isGroup || (contact.isGroup && contact.mucType == "group")) && !HelperTools.isContactBlacklisted(forEncryption:contact.obj)) {
+                if (!contact.isGroup || (contact.isGroup && contact.mucType == "group")) && !HelperTools.isContactBlacklisted(forEncryption:contact.obj) {
                     Button {
-                        if(contact.isEncrypted) {
+                        if contact.isEncrypted {
                             showingShouldDisableEncryptionAlert = true
                         } else {
                             showingCannotEncryptAlert = !contact.obj.toggleEncryption(!contact.isEncrypted)
@@ -132,7 +262,14 @@ struct ContactDetails: View {
                 }
 #endif
                 
-                if(!contact.isGroup && !contact.isSelfChat) {
+                if contact.isGroup && ownAffiliation == "owner" {
+                    let label = contact.obj.mucType == "group" ? NSLocalizedString("Rename Group", comment:"") : NSLocalizedString("Rename Channel", comment:"")
+                    TextField(label, text: $contact.fullNameView, onEditingChanged: {
+                        isEditingNickname = $0
+                    })
+                    .accessibilityLabel(contact.obj.mucType == "group" ? "Group name" : "Channel name")
+                    .addClearButton(isEditing: isEditingNickname, text: $contact.fullNameView)
+                } else if !contact.isGroup && !contact.isSelfChat {
                     TextField(NSLocalizedString("Rename Contact", comment: "placeholder text in contact details"), text: $contact.nickNameView, onEditingChanged: {
                         isEditingNickname = $0
                     })
@@ -148,22 +285,22 @@ struct ContactDetails: View {
                     Text("Pin Chat")
                 }
                 
-                if(contact.obj.isGroup && contact.obj.mucType == "group") {
+                if contact.obj.isGroup && contact.obj.mucType == "group" {
                     NavigationLink(destination: LazyClosureView(MemberList(mucContact:contact))) {
                         Text("Group Members")
                     }
-                } else if(contact.obj.isGroup && contact.obj.mucType == "channel") {
+                } else if contact.obj.isGroup && contact.obj.mucType == "channel" {
                     NavigationLink(destination: LazyClosureView(ChannelMemberList(mucContact:contact))) {
                         Text("Channel Members")
                     }
                 }
 #if !DISABLE_OMEMO
-                if(!HelperTools.isContactBlacklisted(forEncryption:contact.obj)) {
-                    if(!contact.isGroup) {
+                if !HelperTools.isContactBlacklisted(forEncryption:contact.obj) {
+                    if !contact.isGroup {
                         NavigationLink(destination: LazyClosureView(OmemoKeys(contact: contact))) {
                             contact.isSelfChat ? Text("Own Encryption Keys") : Text("Encryption Keys")
                         }
-                    } else if(contact.mucType == "group") {
+                    } else if contact.mucType == "group" {
                         NavigationLink(destination: LazyClosureView(OmemoKeys(contact: contact))) {
                             Text("Encryption Keys")
                         }
@@ -171,7 +308,7 @@ struct ContactDetails: View {
                 }
 #endif
                 
-                if(!contact.isGroup && !contact.isSelfChat) {
+                if !contact.isGroup && !contact.isSelfChat {
                     NavigationLink(destination: LazyClosureView(ContactResources(contact: contact))) {
                         Text("Resources")
                     }
@@ -195,13 +332,13 @@ struct ContactDetails: View {
             Section { // the destructive section...
                 if !contact.isSelfChat {
                     Button(action: {
-                        if(!contact.isBlocked) {
+                        if !contact.isBlocked {
                             showingBlockContactConfirmation = true
                         } else {
                             showingCannotBlockAlert = !contact.obj.toggleBlocked(!contact.isBlocked)
                         }
                     }) {
-                        if(!contact.isBlocked) {
+                        if !contact.isBlocked {
                             Text("Block Contact")
                                 .foregroundColor(.red)
                         } else {
@@ -228,12 +365,12 @@ struct ContactDetails: View {
                     }
 
                     Group {
-                        if(contact.isInRoster) {
+                        if contact.isInRoster {
                             Button(action: {
                                 showingRemoveContactConfirmation = true
                             }) {
-                                if(contact.isGroup) {
-                                    if(contact.mucType == "group") {
+                                if contact.isGroup {
+                                    if contact.mucType == "group" {
                                         Text("Leave Group")
                                             .foregroundColor(.red)
                                     } else {
@@ -265,8 +402,8 @@ struct ContactDetails: View {
                             Button(action: {
                                 showingAddContactConfirmation = true
                             }) {
-                                if(contact.isGroup) {
-                                    if(contact.mucType == "group") {
+                                if contact.isGroup {
+                                    if contact.mucType == "group" {
                                         Text("Join Group")
                                     } else {
                                         Text("Join Channel")
@@ -294,11 +431,54 @@ struct ContactDetails: View {
                     }
                 }
 
+                if ownAffiliation == "owner" {
+                    Section {
+                        Button(action: {
+                            showingDestroyConfirmation = true
+                        }) {
+                            if contact.mucType == "group" {
+                                Text("Destroy Group").foregroundColor(.red)
+                            } else {
+                                Text("Destroy Channel").foregroundColor(.red)
+                            }
+                        }
+                        .actionSheet(isPresented: $showingDestroyConfirmation) {
+                            ActionSheet(
+                                title: contact.mucType == "group" ? Text("Destroy Group") : Text("Destroy Channel"),
+                                message: contact.mucType == "group" ? Text("Do you really want to destroy this group? Every member will be kicked out and it will be destroyed afterwards.") : Text("Do you really want to destroy this channel? Every member will be kicked out and it will be destroyed afterwards."),
+                                buttons: [
+                                    .cancel(),
+                                    .destructive(
+                                        Text("Yes"),
+                                        action: {
+                                            showLoadingOverlay(overlay, headline: contact.mucType == "group" ? NSLocalizedString("Destroying group...", comment: "") : NSLocalizedString("Destroying channel...", comment: ""))
+                                            self.account.mucProcessor.destroyRoom(contact.contactJid as String)
+                                            self.account.mucProcessor.addUIHandler({_data in let data = _data as! NSDictionary
+                                                hideLoadingOverlay(overlay)
+                                                let success : Bool = data["success"] as! Bool;
+                                                if success {
+                                                    if let callback = data["callback"] {
+                                                        self.successCallback = objcCast(callback) as monal_void_block_t
+                                                    }
+                                                    DDLogError("callback: \(String(describing:self.successCallback))")
+                                                    successAlert(title: Text("Success"), message: contact.mucType == "group" ? Text("Successfully destroyed group.") : Text("Successfully destroyed channel."))
+                                                } else {
+                                                    errorAlert(title: Text("Error destroying group!"), message: Text(data["errorMessage"] as! String))
+                                                }
+                                            }, forMuc:contact.contactJid)
+                                        }
+                                    )
+                                ]
+                            )
+                        }
+                    }
+                }
+                
                 Button(action: {
                     showingClearHistoryConfirmation = true
                 }) {
-                    if(contact.isGroup) {
-                        if(contact.obj.mucType == "group") {
+                    if contact.isGroup {
+                        if contact.obj.mucType == "group" {
                             Text("Clear chat history of this group")
                         } else {
                             Text("Clear chat history of this channel")
@@ -329,8 +509,7 @@ struct ContactDetails: View {
             //omemo debug stuff, should be removed in a few months
             Section {
                 // only display omemo session reset button on 1:1 and private groups
-                if(contact.obj.isGroup == false || (contact.isGroup && contact.mucType == "group"))
-                {
+                if contact.obj.isGroup == false || (contact.isGroup && contact.mucType == "group") {
                     Button(action: {
                         showingResetOmemoSessionConfirmation = true
                     }) {
@@ -357,20 +536,28 @@ struct ContactDetails: View {
 #endif
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .addLoadingOverlay(overlay)
         .navigationBarTitle(contact.contactDisplayName as String, displayMode:.inline)
-        .applyClosure { view in
-            if contact.isGroup && isGroupModerator && self.account.accountState.rawValue >= xmppState.stateBound.rawValue {
-                view.toolbar {
-                    ToolbarItem(placement:.navigationBarTrailing) {
-                        let ownAffiliation = DataLayer.sharedInstance().getOwnAffiliation(inGroupOrChannel:contact.obj) ?? "none"
-                        NavigationLink(destination:LazyClosureView(GroupDetailsEdit(contact:contact, ownAffiliation:ownAffiliation))) {
-                            Text("Edit")
-                        }
+        .alert(isPresented: $showAlert) {
+            Alert(title: alertPrompt.title, message: alertPrompt.message, dismissButton:.default(Text("Close"), action: {
+                showAlert = false
+                if self.success == true {
+                    //close muc ui and leave chat ui of this muc
+                    if let callback = self.successCallback {
+                        callback()
+                    }
+                    if let activeChats = (UIApplication.shared.delegate as! MonalAppDelegate).activeChats {
+                        activeChats.presentChat(with:nil)
                     }
                 }
-            } else {
-                view
-            }
+            }))
+        }
+        .onChange(of:inputImage) { _ in
+            showLoadingOverlay(overlay, headline: NSLocalizedString("Uploading image...", comment: ""))
+            self.account.mucProcessor.publishAvatar(inputImage, forMuc: contact.contactJid)
+        }
+        .onChange(of:contact.avatar as UIImage) { _ in
+            hideLoadingOverlay(overlay)
         }
     }
 }
